+++ lib/src/source_visitor.dart
@@ -322,7 +322,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     // If the argument list has a trailing comma, format it like a collection
     // literal where each argument goes on its own line, they are indented +2,
     // and the ")" ends up on its own line.
-    if (node.arguments.last.endToken.next.type == TokenType.COMMA) {
+    if (hasCommaAfter(node.arguments.last)) {
       _visitCollectionLiteral(
           null, node.leftParenthesis, node.arguments, node.rightParenthesis);
       return;
@@ -354,7 +354,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     // If the argument list has a trailing comma, format it like a collection
     // literal where each argument goes on its own line, they are indented +2,
     // and the ")" ends up on its own line.
-    if (arguments.last.endToken.next.type == TokenType.COMMA) {
+    if (hasCommaAfter(arguments.last)) {
       _visitCollectionLiteral(
           null, node.leftParenthesis, arguments, node.rightParenthesis);
       return;
@@ -377,7 +377,7 @@ class SourceVisitor extends ThrowingAstVisitor {
       // If the argument list has a trailing comma, format it like a collection
       // literal where each argument goes on its own line, they are indented +2,
       // and the ")" ends up on its own line.
-      if (arguments.last.endToken.next.type == TokenType.COMMA) {
+      if (hasCommaAfter(arguments.last)) {
         _visitCollectionLiteral(
             null, node.leftParenthesis, arguments, node.rightParenthesis);
         return;
@@ -634,7 +634,7 @@ class SourceVisitor extends ThrowingAstVisitor {
 
     return arguments == null ||
         arguments.arguments.isEmpty ||
-        arguments.arguments.last.endToken.next.type != TokenType.COMMA;
+        !hasCommaAfter(arguments.arguments.last);
   }
 
   /// Whether a cascade should be allowed to be inline as opposed to one
@@ -883,7 +883,7 @@ class SourceVisitor extends ThrowingAstVisitor {
 
   void _visitConstructorInitializers(ConstructorDeclaration node) {
     var hasTrailingComma = node.parameters.parameters.isNotEmpty &&
-        node.parameters.parameters.last.endToken.next.type == TokenType.COMMA;
+        hasCommaAfter(node.parameters.parameters.last);
 
     if (hasTrailingComma) {
       // Since the ")", "])", or "})" on the preceding line doesn't take up
@@ -1070,7 +1070,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     visitCommaSeparatedNodes(node.constants, between: splitOrTwoNewlines);
 
     // If there is a trailing comma, always force the constants to split.
-    if (node.constants.last.endToken.next.type == TokenType.COMMA) {
+    if (hasCommaAfter(node.constants.last)) {
       builder.forceRules();
     }
 
@@ -1212,7 +1212,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     // If the parameter list has a trailing comma, format it like a collection
     // literal where each parameter goes on its own line, they are indented +2,
     // and the ")" ends up on its own line.
-    if (node.parameters.last.endToken.next.type == TokenType.COMMA) {
+    if (hasCommaAfter(node.parameters.last)) {
       _visitTrailingCommaParameterList(node);
       return;
     }
@@ -1248,11 +1248,7 @@ class SourceVisitor extends ThrowingAstVisitor {
 
       for (var param in requiredParams) {
         visit(param);
-
-        // Write the following comma.
-        if (param.endToken.next.type == TokenType.COMMA) {
-          token(param.endToken.next);
-        }
+        _writeCommaAfter(param);
 
         if (param != requiredParams.last) rule.beforeArgument(split());
       }
@@ -1279,11 +1275,7 @@ class SourceVisitor extends ThrowingAstVisitor {
 
       for (var param in optionalParams) {
         visit(param);
-
-        // Write the following comma.
-        if (param.endToken.next.type == TokenType.COMMA) {
-          token(param.endToken.next);
-        }
+        _writeCommaAfter(param);
 
         if (param != optionalParams.last) namedRule.beforeArgument(split());
       }
@@ -2822,11 +2814,7 @@ class SourceVisitor extends ThrowingAstVisitor {
       }
 
       visit(element);
-
-      // The comma after the element.
-      if (element.endToken.next.type == TokenType.COMMA) {
-        token(element.endToken.next);
-      }
+      _writeCommaAfter(element);
     }
 
     builder.endRule();
@@ -2835,10 +2823,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     var force = _collectionSplits.removeLast();
 
     // If the collection has a trailing comma, the user must want it to split.
-    if (elements.isNotEmpty &&
-        elements.last.endToken.next.type == TokenType.COMMA) {
-      force = true;
-    }
+    if (elements.isNotEmpty && hasCommaAfter(elements.last)) force = true;
 
     if (node != null) _endPossibleConstContext(node.constKeyword);
     _endLiteralBody(rightBracket, ignoredRule: rule, forceSplit: force);
@@ -2884,11 +2869,7 @@ class SourceVisitor extends ThrowingAstVisitor {
 
     for (var parameter in parameters.parameters) {
       visit(parameter);
-
-      // The comma after the parameter.
-      if (parameter.endToken.next.type == TokenType.COMMA) {
-        token(parameter.endToken.next);
-      }
+      _writeCommaAfter(parameter);
 
       // If the optional parameters start after this one, put the delimiter
       // at the end of its line.
@@ -2991,8 +2972,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     var parent = node.parent;
     if (parent is NamedExpression) parent = parent.parent;
 
-    return parent is ArgumentList &&
-        parent.arguments.last.endToken.next.type == TokenType.COMMA;
+    return parent is ArgumentList && hasCommaAfter(parent.arguments.last);
   }
 
   /// Whether [node] is a spread of a collection literal.
@@ -3263,6 +3243,30 @@ class SourceVisitor extends ThrowingAstVisitor {
       _writeText(line, offset);
       offset += line.length;
     }
+  }
+
+  /// Write the comma token following [node], if there is one.
+  void _writeCommaAfter(AstNode node) {
+    token(_commaAfter(node));
+  }
+
+  /// Whether there is a comma token immediately following [node].
+  bool hasCommaAfter(AstNode node) => _commaAfter(node) != null;
+
+  /// The comma token immediately following [node] if there is one, or `null`.
+  Token _commaAfter(AstNode node) {
+    if (node.endToken.next.type == TokenType.COMMA) {
+      return node.endToken.next;
+    }
+
+    // TODO(sdk#38990): endToken doesn't include the "?" on a nullable
+    // function-typed formal, so check for that case and handle it.
+    if (node.endToken.next.type == TokenType.QUESTION &&
+        node.endToken.next.next.type == TokenType.COMMA) {
+      return node.endToken.next.next;
+    }
+
+    return null;
   }
 
   /// Emit the given [modifier] if it's non null, followed by non-breaking